+++ packages/thirdweb/src/exports/react.ts
@@ -113,6 +113,7 @@ export {
   AutoConnect,
   type AutoConnectProps,
 } from "../react/core/hooks/connection/AutoConnect.js";
+export { useAutoConnect } from "../react/core/hooks/connection/useAutoConnect.js";
 
 // auth
 export { type SiweAuthOptions } from "../react/core/hooks/auth/useSiweAuth.js";